Moved makeIdFromString to Core Tracker - created angular filter for it
