+++ tracker/trackers/angular/src/index.ts
@@ -3,5 +3,7 @@
  */
 
 export * from './objectiv-tracker.directive';
+export * from './objectiv-tracker.filter';
+export * from './objectiv-tracker.initializer';
 export * from './objectiv-tracker.module';
 export * from './objectiv-tracker.token';
