+++ tracker/trackers/react/src/common/factories/makeTitleFromChildren.ts
@@ -8,8 +8,6 @@ import { recursiveGetTextFromChildrenNode } from './recursiveGetTextFromChildren
 /**
  * Retrieve text from given ReactNode children.
  * The resulting text may be used, among others, to infer a valid text and identifier for a Button.
- *
- * @see makeIdFromString
  */
 export const makeTitleFromChildren = (children: ReactNode): string => {
   return recursiveGetTextFromChildrenNode(children)?.trim() ?? '';
