+++ tracker/trackers/react/src/index.ts
@@ -7,7 +7,6 @@ export * from './common/factories/makeDefaultPluginsList';
 export * from './common/factories/makeDefaultQueue';
 export * from './common/factories/makeDefaultTransport';
 export * from './common/factories/makeExpandableContext';
-export * from './common/factories/makeIdFromString';
 export * from './common/factories/makeInputContext';
 export * from './common/factories/makeLinkContext';
 export * from './common/factories/makeLocationContext';
